Move the school list item to shared component
