+++ app/screens/MajorSelection/MajorRecommendationScreen.js
@@ -6,12 +6,19 @@ import { StackActions } from '@react-navigation/native';
 import BoldLabelComponent from '../../components/shared/BoldLabelComponent';
 import ConfirmationModal from '../../components/shared/ConfirmationModal';
 import CustomNavigationHeader from '../../components/shared/CustomNavigationHeader';
+import SchoolList from '../../components/schools/school_list';
 import Quiz from '../../models/Quiz';
+import {screenHorizontalPadding} from '../../constants/component_constant';
+import { FontSetting } from '../../assets/style_sheets/font_setting';
+import {FontFamily} from '../../themes/font';
+import School from '../../models/School';
 
 const MajorRecommendationScreen = ({route, navigation}) => {
   const [modalVisible, setModalVisible] = React.useState(false);
   const currentQuiz = Quiz.findByUuid(route.params.quizUuid);
   const major = currentQuiz.selectedMajor || {};
+
+  console.log('== Major = ', major)
 
   useEffect(() => {
     const backHandler = BackHandler.addEventListener('hardwareBackPress', () => {
@@ -43,12 +50,28 @@ const MajorRecommendationScreen = ({route, navigation}) => {
     navigation.dispatch(StackActions.replace('HollandNavigator'))
   }
 
+  const renderTitle = (label) => {
+    return <Text style={{marginLeft: screenHorizontalPadding, fontSize: FontSetting.title, marginBottom: 6, marginTop: 16, fontFamily: FontFamily.bold}}>{label}</Text>
+    // return <Text style={{marginLeft: screenHorizontalPadding, fontSize: FontSetting.title, marginBottom: 6, marginTop: 16, fontFamily: FontFamily.regular}}>{label}</Text>
+  }
+
+  const renderRelatedSchools = () => {
+    const schools = major.school_ids.map(schoolId => School.findById(schoolId));
+    return (
+      <View>
+        {renderTitle('គ្រឺះស្ថានសិក្សា')}
+        <View>
+          <SchoolList navigation={navigation} data={schools} />
+        </View>
+      </View>
+    )
+  }
+
   return (
     <View style={{flex: 1}}>
       <CustomNavigationHeader title='ជម្រើសជំនាញកម្រិតឧត្តមសិក្សា' headerStyle={{zIndex: 1}} onPressBack={() => navigation.dispatch(StackActions.replace('HollandNavigator'))} />
       <ScrollView>
-        <Text style={{textAlign: 'center', marginVertical: 16}}>ការផ្តល់អនុសាសន៍</Text>
-
+        {renderTitle('ការផ្តល់អនុសាសន៍')}
         <Card style={{padding: 16}}>
           <Text>
             ដើម្បីអាចបន្តការសិក្សាលើមុខជំនាញ.."<BoldLabelComponent label={major.name}/>".. នៅកម្រិតឧត្តមសិក្សាទទួល បានជោគជ័យ ប្អូនត្រូវពិនិត្យលើចំណុចមួយ ចំនួនដូច ខាងក្រោម៖
@@ -56,6 +79,9 @@ const MajorRecommendationScreen = ({route, navigation}) => {
             { major.recommendation }
           </Text>
         </Card>
+
+
+        {renderRelatedSchools()}
 
       </ScrollView>
       {renderModal()}
